+++ src/config/userRules.js
@@ -16,7 +16,8 @@ export const rulesList = {
   'pdr_view': 'sales:view',
   'financial_view': 'financial:view',
   'upload_statement': 'upload:statement',
-  'projection_report': 'projection:report'
+  'projection_report': 'projection:report',
+  'applicant_delete': 'applicant:delete'
 }
 
 export const rulesForUserRoles = {
@@ -70,6 +71,7 @@ export const rulesForUserRoles = {
       rulesList.dealer_credit_view,
       rulesList.dealership_credit_view,
       rulesList.credit_view,
+      rulesList.pdr_view,
       rulesList.financial_view,
       rulesList.upload_statement,
     ]
@@ -90,6 +92,7 @@ export const rulesForUserRoles = {
       rulesList.financial_view,
       rulesList.upload_statement,
       rulesList.projection_report,
+      rulesList.applicant_delete
     ]
   },
   CREDIT_EXEC: {
@@ -101,8 +104,10 @@ export const rulesForUserRoles = {
       rulesList.dealer_credit_edit,
       rulesList.dealership_credit_edit,
       rulesList.credit_view,
+      rulesList.pdr_view,
       rulesList.financial_view,
       rulesList.upload_statement,
+      rulesList.region_map,
     ]
   },
   OPS_EXEC: {
@@ -135,8 +140,10 @@ export const rulesForUserRoles = {
       rulesList.dealer_credit_edit,
       rulesList.dealership_credit_edit,
       rulesList.credit_view,
+      rulesList.pdr_view,
       rulesList.financial_view,
       rulesList.upload_statement,
+      rulesList.region_map,
     ]
   },
   SALES_HEAD_STATE: {