+++ src/pages/transports/components/AddNewVehicleForm.js
@@ -12,8 +12,8 @@ import { useSnackbar } from 'notistack';
 import React, { useState } from 'react';
 import { useQueryClient } from 'react-query';
 import * as Yup from 'yup';
+import { VehicleDetails } from './VehicleDetails';
 import Button from '../../../components/CommonComponents/Button/Button';
-import { ViewData } from '../../../components/CommonComponents/FilePreview';
 import TextInput from '../../../components/TextInput/TextInput';
 import {
   addNewVehicle,
@@ -22,7 +22,6 @@ import {
 
 const useStyles = makeStyles((theme) => ({
   sidePanelTitle: {
-    // textAlign: 'center',
     padding: '24px 16px',
     display: 'flex',
     justifyContent: 'space-between',
@@ -113,7 +112,7 @@ const AddNewVehicleForm = ({
           });
           getVehicleInfoFromID(id)
             .then((data) => {
-              setVehicleDetails(JSON.parse(data?.find(item => item.tt_no === formData.tt_no)?.vehicle_details))
+              setVehicleDetails(data?.find(item => item.tt_no === formData.tt_no)?.vehicle_details)
             })
             .catch(e => console.log(e))
         })
@@ -178,80 +177,7 @@ const AddNewVehicleForm = ({
             </form>
             {
               vehicleDetails &&
-                <Grid container spacing={2} style={{marginTop: 15}}>
-                  <Grid item md={6} className={classes.items}>
-                    <ViewData title="Vehicle Description" value={vehicleDetails?.vehicleClassDescription} style={{marginBottom: 0}} />
-                  </Grid>
-                  <Grid item md={6} className={classes.items}>
-                    <ViewData title="Registration Date" value={vehicleDetails?.registrationDate} style={{marginBottom: 0}} />
-                  </Grid>
-                  <Grid item md={6} className={classes.items}>
-                    <ViewData title="Vehicle Category" value={vehicleDetails?.vehicleCatgory} style={{marginBottom: 0}} />
-                  </Grid>
-                  <Grid item md={6} className={classes.items}>
-                    <ViewData title="Registration No" value={vehicleDetails?.registrationNumber} style={{marginBottom: 0}} />
-                  </Grid>
-                  <Grid item md={6} className={classes.items}>
-                    <ViewData title="Engine No" value={vehicleDetails?.engineNumber} style={{marginBottom: 0}} />
-                  </Grid>
-                  <Grid item md={6} className={classes.items}>
-                    <ViewData title="Chassis No" value={vehicleDetails?.chassisNumber} style={{marginBottom: 0}} />
-                  </Grid>
-                  <Grid item md={6} className={classes.items}>
-                    <ViewData title="No.of Cylinders" value={vehicleDetails?.numberOfCylinders} style={{marginBottom: 0}} />
-                  </Grid>
-                  <Grid item md={6} className={classes.items}>
-                    <ViewData title="Maker Description" value={vehicleDetails?.makerDescription} style={{marginBottom: 0}} />
-                  </Grid>
-                  <Grid item md={6} className={classes.items}>
-                    <ViewData title="Fuel Type" value={vehicleDetails?.fuelDescription} style={{marginBottom: 0}} />
-                  </Grid>
-                  <Grid item md={6} className={classes.items}>
-                    <ViewData title="Model" value={vehicleDetails?.makerModel} style={{marginBottom: 0}} />
-                  </Grid>
-                  <Grid item md={6} className={classes.items}>
-                    <ViewData title="Capacity" value={vehicleDetails?.cubicCapacity} style={{marginBottom: 0}} />
-                  </Grid>
-                  <Grid item md={6} className={classes.items}>
-                    <ViewData title="Color" value={vehicleDetails?.color} style={{marginBottom: 0}} />
-                  </Grid>
-                  <Grid item md={6} className={classes.items}>
-                    <ViewData title="Owner Name" value={vehicleDetails?.ownerName} style={{marginBottom: 0}} />
-                  </Grid>
-                  <Grid item md={6} className={classes.items}>
-                    <ViewData title="Insurance Upto" value={vehicleDetails?.insuranceUpto} style={{marginBottom: 0}} />
-                  </Grid>
-                  <Grid item md={6} className={classes.items}>
-                    <ViewData title="Insurance Policy No" value={vehicleDetails?.insurancePolicyNumber} style={{marginBottom: 0}} />
-                  </Grid>
-                  <Grid item md={6} className={classes.items}>
-                    <ViewData title="Fitness Upto" value={vehicleDetails?.fitnessUpto} style={{marginBottom: 0}} />
-                  </Grid>
-                  <Grid item md={6} className={classes.items}>
-                    <ViewData title="Manufactured Month Year" value={vehicleDetails?.manufacturedMonthYear} style={{marginBottom: 0}} />
-                  </Grid>
-                  <Grid item md={6} className={classes.items}>
-                    <ViewData title="Insurance Company" value={vehicleDetails?.insuranceCompany} style={{marginBottom: 0}} />
-                  </Grid>
-                  <Grid item md={6} className={classes.items}>
-                    <ViewData title="PUC No" value={vehicleDetails?.pucNumber} style={{marginBottom: 0}} />
-                  </Grid>
-                  <Grid item md={6} className={classes.items}>
-                    <ViewData title="PUC Exp Date" value={vehicleDetails?.pucExpiryDate} style={{marginBottom: 0}} />
-                  </Grid>
-                  <Grid item md={6} className={classes.items}>
-                    <ViewData title="Blacklist Status" value={vehicleDetails?.blackListStatus} style={{marginBottom: 0}} />
-                  </Grid>
-                  <Grid item md={6} className={classes.items}>
-                    <ViewData title="Rc Status" value={vehicleDetails?.rcStatus} style={{marginBottom: 0}} />
-                  </Grid>
-                  <Grid item md={6} className={classes.items}>
-                    <ViewData title="Rc Mob No" value={vehicleDetails?.rcMobileNo} style={{marginBottom: 0}} />
-                  </Grid>
-                  <Grid item md={6} className={classes.items}>
-                    <ViewData title="Blacklist Info" value={vehicleDetails?.blackListInfo} style={{marginBottom: 0}} />
-                  </Grid>
-                </Grid>
+                <VehicleDetails vehicleTestDet={vehicleDetails} />
             }
           </Box>
           <Button